[Pallas] Make rng tests inherit from JaxTestCase
PiperOrigin-RevId: 839963758

+++ jax/experimental/pallas/ops/tpu/random/philox.py
@@ -117,7 +117,7 @@ def philox_4x32_kernel(key,
     offset = prng_utils.compute_scalar_offset(
         counts_idx, unpadded_shape, block_shape)
     counts_lo = prng_utils.blocked_iota(block_size, unpadded_shape)
-    counts_lo = counts_lo + offset + offset_ref[0]
+    counts_lo = counts_lo + offset.astype(jnp.uint32) + offset_ref[0]
     counts_lo = counts_lo.astype(jnp.uint32)
     # TODO(justinfu): Support hi bits on count.
     _zeros = jnp.zeros_like(counts_lo)
